Finished raise command, new bugs, however...
* Raising command now completely works, however there is a new bug
which allows the challenging player to accept the challenge
* Added more get methods to match data
* minor changes in acceptTimer scheduler
* Major changes in challenge cmd, allows command to be used by either
player to raise, or accept/decline. It needs some sort of way to make
sure players dont use it out of turn. (should volley back and forth)
* Added set combat wager method to wagerHandler. This is a temporary
storage method for when the players are deciding (raising) what the the
match wager for the participants should be.